+++ {{cookiecutter.repo_name}}/{{cookiecutter.repo_name}}/settings/production.py
@@ -2,25 +2,25 @@ from settings.staging import *
 
 
 # Allowed hosts for the site
-ALLOWED_HOSTS = ['{{ cookiecutter.live_hostname }}']
+ALLOWED_HOSTS = env.list('DJANGO_ALLOWED_HOSTS', default=['{{ cookiecutter.live_hostname }}'])
 
 # Static site url, used when we need absolute url but lack request object, e.g. in email sending.
-SITE_URL = 'https://{{ cookiecutter.live_hostname }}'
+SITE_URL = env.str('DJANGO_SITE_URL', default='https://{{ cookiecutter.live_hostname }}')
 
-EMAIL_HOST_PASSWORD = 'TODO (api key)'
+EMAIL_HOST_PASSWORD = env.str('DJANGO_EMAIL_HOST_PASSWORD', default='TODO (api key)')
 
-RAVEN_BACKEND_DSN = 'https://TODO:TODO@sentry.thorgate.eu/TODO'
-RAVEN_PUBLIC_DSN = 'https://TODO@sentry.thorgate.eu/TODO'
+RAVEN_BACKEND_DSN = env.str('DJANGO_RAVEN_BACKEND_DSN', default='https://TODO:TODO@sentry.thorgate.eu/TODO')
+RAVEN_PUBLIC_DSN = env.str('DJANGO_RAVEN_PUBLIC_DSN', default='https://TODO@sentry.thorgate.eu/TODO')
 RAVEN_CONFIG['dsn'] = RAVEN_BACKEND_DSN
 
 # Enable {{ cookiecutter.django_media_engine }} storage
 DEFAULT_FILE_STORAGE = '{{ cookiecutter.repo_name }}.storages.MediaStorage'
-MEDIA_ROOT = ''
+MEDIA_ROOT = env.str('DJANGO_MEDIA_ROOT', default='')
 {% if cookiecutter.django_media_engine == 'S3' -%}
-AWS_STORAGE_BUCKET_NAME = '{{ cookiecutter.repo_name }}-production'
-AWS_ACCESS_KEY_ID = None  # set in local.py
-AWS_SECRET_ACCESS_KEY = None  # set in local.py
+AWS_STORAGE_BUCKET_NAME = env.str('DJANGO_AWS_STORAGE_BUCKET_NAME', default='{{ cookiecutter.repo_name }}-production')
+AWS_ACCESS_KEY_ID = env.str('DJANGO_AWS_ACCESS_KEY_ID')
+AWS_SECRET_ACCESS_KEY = env.str('DJANGO_AWS_SECRET_ACCESS_KEY')
 {%- endif %}{% if cookiecutter.django_media_engine == 'GCS' -%}
-GS_BUCKET_NAME = '{{ cookiecutter.repo_name }}-production'
-GS_PROJECT_ID = None  # set in local.py
-GS_CREDENTIALS = None  # set in local.py{% endif %}
+GS_BUCKET_NAME = env.str('DJANGO_GS_BUCKET_NAME', default='{{ cookiecutter.repo_name }}-production')
+GS_PROJECT_ID = env.str('DJANGO_GS_PROJECT_ID')
+GS_CREDENTIALS = env.str('DJANGO_GS_CREDENTIALS'){% endif %}